Add commit template with commit msg hints
The hints remind me to write a bit more into the commit message than
just a short title.

An alternative would be a prepare-commit hook, but those are often used
by projects. The template I can check into my personal git settings and
overwrite it in specific projects if needed.